split up prisma func
