+++ src/cdk/lambda/node14func.ts
@@ -1,10 +1,7 @@
-/* eslint-disable prefer-const */
 import "source-map-support/register"
 import { BundlingOptions, NodejsFunction, NodejsFunctionProps } from "@aws-cdk/aws-lambda-nodejs"
 import { Construct } from "@aws-cdk/core"
 import { Runtime } from "@aws-cdk/aws-lambda"
-import { Vpc } from "@aws-cdk/aws-ec2"
-import { DB_SECRET_ENV, SlsPgDb } from "../.."
 
 export type Node14FuncProps = NodejsFunctionProps
 
@@ -46,40 +43,5 @@ export class Node14Func extends NodejsFunction {
     super(scope, id, newProps)
 
     this.bundling = bundling
-  }
-}
-
-export interface DatabaseFuncProps extends Node14FuncProps {
-  vpc: Vpc
-  db: SlsPgDb
-}
-
-/**
- * Lambda function with Prisma layer and generated client.
- * Grants access to DB and secret.
- * @alpha
- */
-export class PrismaNode14Func extends Node14Func {
-  constructor(scope: Construct, id: string, { db, bundling, layers, ...props }: DatabaseFuncProps) {
-    // add prisma layer
-    layers ||= []
-    bundling ||= {}
-    ;(bundling as any).externalModules ||= []
-    bundling.externalModules!.push(...db.getPrismaExternalModules())
-    layers.push(db.getPrismaLayerVersion())
-
-    super(scope, id, {
-      ...props,
-      layers,
-      bundling,
-    })
-
-    // allow DB access
-    db.grantDataApiAccess(this)
-    db.connections.allowDefaultPortFrom(this)
-    if (db.secret) {
-      this.addEnvironment(DB_SECRET_ENV, db.secret.secretArn)
-      db.secret.grantRead(this)
-    }
   }
 }
